Separe benchmarks from tests
