+++ benchmarks/data_files/data/grids/208/HERA_CC_318GEV_EM_SIGMARED.pineappl.lz4
