+++ benchmarks/data_files/data/operator_cards/_template.yaml
