+++ benchmarks/theory_bench.py
