+++ benchmarks/theory_card_bench.py
@@ -3,7 +3,7 @@ import pathlib
 
 import pineko
 
-test_files = pathlib.Path(__file__).parents[0] / "test_files/"
+test_files = pathlib.Path(__file__).parents[0] / "data_files/"
 
 
 def test_load():
