--- tests/test_files/pineko.toml
@@ -1,14 +1,0 @@
-[paths]
-# inputs
-ymldb = "data/yamldb"
-grids = "data/grids"
-theory_cards = "data/theory_cards"
-operator_card_template = "data/operator_cards/_template.yaml"
-# outputs
-operator_cards = "data/operator_cards"
-ekos = "data/ekos"
-fktables = "data/fktables"
-
-[paths.logs]
-fk = "logs/fk"
-eko = "logs/eko"
